+++ lib/err/sub-pkg/err.ml
@@ -1,7 +1,0 @@
-(*********************************************************************************)
-(*  pplumbing - Utility libraries to use with [pp]                               *)
-(*  SPDX-FileCopyrightText: 2024-2025 Mathieu Barbin <mathieu.barbin@gmail.com>  *)
-(*  SPDX-License-Identifier: MIT                                                 *)
-(*********************************************************************************)
-
-include Pplumbing_err.Err